+++ pkg/kiosk/local_login.go
@@ -22,7 +22,6 @@ func GrafanaKioskLocal(cfg *Config) {
 	opts := []chromedp.ExecAllocatorOption{
 		chromedp.NoFirstRun,
 		chromedp.NoDefaultBrowserCheck,
-		// chromedp.DisableGPU, // needed?
 		chromedp.Flag("noerrdialogs", true),
 		chromedp.Flag("kiosk", true),
 		chromedp.Flag("bwsi", true),
@@ -32,6 +31,8 @@ func GrafanaKioskLocal(cfg *Config) {
 		chromedp.Flag("disable-overlay-scrollbar", true),
 		chromedp.Flag("ignore-certificate-errors", cfg.Target.IgnoreCertificateErrors),
 		chromedp.Flag("test-type", cfg.Target.IgnoreCertificateErrors),
+		chromedp.Flag("window-position", cfg.General.WindowPosition),
+		chromedp.Flag("check-for-update-interval", 31536000),
 		chromedp.UserDataDir(dir),
 	}